Casted integer to double because of problems with c++ implementations.

+++ OpenMPvalidation/SRC/V2/c/omp_for_private.c
@@ -13,7 +13,7 @@ static void do_some_work (){
     int i;
     double sum = 0;
     for(i = 0; i < 1000; i++){
-	sum += sqrt (i);
+	sum += sqrt ((double) i);
     }
 }
